feat: interface ITemplateBiometriaReadOnlyRepository em Domain

+++ src/Backend/ControleDeAcesso26.Domain/Interfaces/ITemplateBiometriaUsuario/ITemplateBiometriaReadOnlyRepository.cs
@@ -4,7 +4,7 @@ namespace ControleDeAcesso26.Domain.Interfaces.ITemplateBiometriaUsuario
 {
     public interface ITemplateBiometriaReadOnlyRepository
     {
-        public Task<bool> IdSensor1JaExiste(long idSensor);
+        public Task<bool> IdSensor1JaExiste(int idSensor);
         public Task<TemplateBiometriaUsuario?> BuscarTemplatePorId(int idSensor);
     }
 }
